Formatting - run the new formatting against the Brain and MidiInputTransformer files

+++ Source/MidiGenerator/MidiInputTransformer/MidiInputTransformer.cpp
@@ -10,13 +10,15 @@
 #include <stdexcept>
 #include <utility>
 
-std::vector<ModelVector>
-MidiInputTransformer::getModelInputForBuffer(MidiBuffer &buffer) {
+std::vector<ModelVector> MidiInputTransformer::getModelInputForBuffer(
+  MidiBuffer &buffer)
+{
   return parseMidiInput(buffer);
 }
 
-std::vector<ModelVector>
-MidiInputTransformer::parseMidiInput(MidiBuffer &buffer) {
+std::vector<ModelVector> MidiInputTransformer::parseMidiInput(
+  MidiBuffer &buffer)
+{
   using InputMap = std::map<int, ModelVector>;
   InputMap inputBySample;
   int sampleNumber;
@@ -25,11 +27,11 @@ MidiInputTransformer::parseMidiInput(MidiBuffer &buffer) {
     if (message.isNoteOn()) {
       if (inputBySample.count(sampleNumber) == 1) {
         inputBySample.at(sampleNumber).data.at(message.getNoteNumber()) +=
-            message.getFloatVelocity();
+          message.getFloatVelocity();
       } else {
-        inputBySample.emplace(
-            std::make_pair(sampleNumber,
-                           midiMessageToModelInput(message, sampleNumber)));
+        inputBySample.emplace(std::make_pair(
+          sampleNumber,
+          midiMessageToModelInput(message, sampleNumber)));
       }
     }
   }
@@ -41,15 +43,17 @@ MidiInputTransformer::parseMidiInput(MidiBuffer &buffer) {
   return transformedInput;
 }
 
-ModelVector
-MidiInputTransformer::midiMessageToModelInput(const MidiMessage &message,
-                                              int sampleNumber) {
+ModelVector MidiInputTransformer::midiMessageToModelInput(
+  const MidiMessage &message,
+  int sampleNumber)
+{
   if (!message.isNoteOn()) {
     throw std::invalid_argument("MIDI message is not noteOn");
   }
   ModelVectorData data{};
   data.at(message.getNoteNumber()) = message.getFloatVelocity();
-  return ModelVector{data,
-                     (int64_t)sampleNumber,
-                     ModelVector::InputType::MidiInput};
+  return ModelVector{
+    data,
+    (int32_t)sampleNumber,
+    ModelVector::InputType::MidiInput};
 }
